// src/models/Events.enum.ts
export enum Events {
  LOGIN_ATTEMPT = 'LOGIN_ATTEMPT',
  LOGIN_RESPONSE = 'LOGIN_RESPONSE',
  REFRESH_USERS = 'REFRESH_USERS',
  INVITATION_SENT = 'INVITATION_SENT',
  INVITATION_GOT = 'INVITATION_GOT',
  INVITATION_ACCEPT = 'INVITATION_ACCEPT',
  CHATROOM_CREATE = 'CHATROOM_CREATE',
  CHAT_MESSAGE = 'CHAT_MESSAGE',
  CHAT_LEAVE = 'CHAT_LEAVE',
}